DBdust expansion and DBshow bug fix
DBdust pointers have been expanded to 8 byte pointers to accommodate
large data sets.  A patch program DUSTupgrade.Jan.1.2015 has been added.

DBshow has been fixed to produce correct fast headers on database
blocks (it was correct for an entire DB).

+++ DBstats.c
@@ -239,11 +239,11 @@ int main(int argc, char *argv[])
         dust = Load_Track(&db,Numbered_Suffix("",db.part,".dust")); 
       }
     if (dust != NULL)
-      { char *data = dust->data;
-        int  *anno = (int *) dust->anno;
-        int   i, rlen;
-        int  *idata, *edata;
-        int64 numint, totlen;
+      { char  *data = dust->data;
+        int64 *anno = (int64 *) dust->anno;
+        int    i, rlen;
+        int   *idata, *edata;
+        int64  numint, totlen;
 
         totlen = 0;
         numint = 0;
